exercises/01-sigstore/requirements.txt
requests==2.28.1
requests-oauthlib==1.3.1
cryptography==37.0.4
pyjwt==2.4.0